refactor(ui): modular component structure with shared utilities
- Split api/client.ts into domain modules: fires, archive, forecast, review, watchlist, http
- Split FireDetailsPanel (1068 lines) into fire-details/ subfolder with tab components, ForecastPanel, QAReviewPanel, NewsCard, NewsExpandedModal
- Split FireMap (808 lines) into components/map/ with layerConfig constants and MapTooltip
- Extract hooks: useAppDerivedState, useMapInteractions
- Extract components: ArchiveControls, AppLayout
- Slim App.tsx from 864 → ~336 lines
- Consolidate 4 duplicate HIGH_CONFIDENCE_THRESHOLD, 4x geometryProvenanceLabel, 4x toFiniteNumber/safeNumber, 2x TIMEFRAME_ICONS, 2x isoFormat into single exports
- Move SAFETY_TIER_COLORS and StatCard type out of render scope
- Remove unnecessary visibleEvents shallow copy in FireMap
